Fix loop import
